Modified vaporization code to only give a chance at vaporizing.
If the item is non-limited, never vaporizes.
It the item is limited, but under (or at) limit, it never vaporizes.
If the item is limited and over limit, there is a chance to vaporize
which is higher based on the amount over limit and the rareness of
the item.  So limit 1 with 2 copies, is 50% chance to vaporize.  Limit
2 with 3 copies is 33% chance to vaporize.  Limit 10 with 11 copies
is 9% chance to vaporize... etc.