Add some file transfer utility and OLC niceness
Adds a localxfer folder that is mounted since we can't mount the
area folder directly yet (One day). Updated readme with commands for
running copy command on the container to pull an area file off the
container and into localxfer as a way to pull the copies.

Updated OLC to show the `done` command that lets you out of the
OLC editor.

Updated the deployments to add the local path mount for windows for
localxfer

Added localxfer ignore to build for tilt so it doesn't rebuild when
you copy something out.